shorter

+++ handler/db/db.go
@@ -4,5 +4,5 @@ type Database interface {
 	Close() error
 	Set(index, shortURL string) error
 	Get(index string) (string, error)
-	Len() (uint64, error)
+	Len() (int, error)
 }
